UI : Moving Sample data to its separate tab.
